Basic implementation of collection of filenames
This is the first basic implementation of main calling a function that
gathers the filenames of the target files. This does not use the struct
that we want to use yet, but I learned a lot about dynamic memory here.
I even used valgrind to fix memory leaks from not freeing stuff
afterwards. This version has no memory leaks.